backend/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
python-dotenv==1.0.0
python-docx==0.8.11
beautifulsoup4==4.12.0
requests==2.31.0
jieba==0.42.1
reportlab==4.0.0
markdown==3.5.1
PyMuPDF==1.23.8
pdfplumber==0.10.3
Pillow==10.1.0
